Dates support for Course student page  Changes to be committed: 	modified:   school-service/src/main/java/school/service/CourseService.java 	modified:   school-service/src/main/java/school/service/implementation/CourseServiceImpl.java 	modified:   school-web/src/main/java/school/controller/CourseController.java 	modified:   school-web/src/main/webapp/views/courses/student.jsp

+++ school-service/src/main/java/school/service/CourseService.java
@@ -6,5 +6,5 @@ import java.util.List;
 import school.model.Course;
 
 public interface CourseService {
-    List<Course> getCourseByUserIdAndDataRange(int id, Date from, Date till);
+    List<Course> getCourseByUserIdAndDataRange(long id, Date from, Date till);
 }
